Give output in the correct format (unsorted)
Entries are not topologically sorted so the installation order may be invalid.
Furthermore, a non-cyclic sort may not be possible. This could be due to:

1) The "bootstrapping" problem, a la Seen 6, in which case there is a missing
package which must temporarily be installed, then uninstalled before the final
state.

2) The final state found by the solver may be unreachable. In which case the
program should return to the solver, adding the cost of the unreachable state as
a lower bound on the cost of the new state.