Refactor use of second prover
Valid solutions are now correctly ordered. There are still invalid solutions
generated e.g.

Seen 3: ["+A=1","+B=1","+D=1"]
Invalid because while this stat is consistent, it is not reachable from the
initial state without installing and then uninstalling "C=1". The obvious
indicator that this solution is wrong is that the final step should be "+A=1".

Seen 6: ["+X2=0","+X1=1","+X3=0","+X5=0","+X4=1"]
This is the "Bootstrapping" problem, which is not solvable in a single-step
PseudoBoolean approach.